Move component and test into their own dir
